+++ src/FalconCmsServiceProvider.php
@@ -60,6 +60,7 @@ class FalconCmsServiceProvider extends ServiceProvider
             \FalconCms\Core\Console\Commands\FalconList::class,
             \FalconCms\Core\Console\Commands\MakeDashboardPage::class,
             \FalconCms\Core\Console\Commands\InstallFalconCms::class,
+            \FalconCms\Core\Console\Commands\UninstallFalconCms::class,
             \FalconCms\Core\Console\Commands\SeedFalconCms::class,
             \FalconCms\Core\Console\Commands\UpdateFalconCms::class,
             \FalconCms\Core\Console\Commands\PublishScheduledPosts::class,